third_party/blink/web_tests/fast/lists/list-color-change-no-layout.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
onload = function() {
  document.body.offsetTop;
  var target = document.getElementById('target');
  target.style.color = 'green';
  if (window.internals) {
    shouldBe('internals.updateStyleAndReturnAffectedElementCount()', '3');
    shouldBe('internals.needsLayoutCount()', '0');
  }
}
</script>
Should not layout when list color changes.
<ul id="target">
  <li>List item</li>
</ul>
