third_party/blink/web_tests/fast/lists/list-color-change-no-layout-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS internals.needsLayoutCount() is 0
Should not layout when list color changes.
List item
